Rename project to AppImageInfo
