+++ include/appimage/appimageinfo.h
