+++ src/CMakeLists.txt
@@ -1,5 +1,5 @@
 
-add_library(libappimage-info SHARED
+add_library(libappimageinfo SHARED
         entities/FileMetadataExtractor.cpp
         entities/FileMetadataExtractor.h
         entities/DesktopFileMetadataExtractor.cpp
@@ -10,13 +10,12 @@ add_library(libappimage-info SHARED
         entities/FileMetadataMerger.h
         entities/BinaryMetadataExtractor.cpp
         entities/BinaryMetadataExtractor.h
-        ${PROJECT_SOURCE_DIR}/include/appimage/appimage-info.h
-        lib/appimage-info-impl.cpp)
+        ${PROJECT_SOURCE_DIR}/include/appimage/appimageinfo.h
+        lib/appimageinfo.cpp)
 
 
-set_target_properties(libappimage-info PROPERTIES PUBLIC_HEADER ${PROJECT_SOURCE_DIR}/include/appimage/appimage-info.h)
-target_compile_options(libappimage-info PUBLIC -fPIC -fexceptions)
-target_include_directories(libappimage-info
+target_compile_options(libappimageinfo PUBLIC -fPIC -fexceptions)
+target_include_directories(libappimageinfo
         PUBLIC
         ${PROJECT_SOURCE_DIR}/include
         PRIVATE
@@ -26,22 +25,16 @@ target_include_directories(libappimage-info
         ${OPENSSL_INCLUDE_DIRS}
         ${NLOHMANN_JSON_INCLUDE_DIRS})
 
-target_link_libraries(libappimage-info
+target_link_libraries(libappimageinfo
         ${LIBAPPIMAGE_LIBRARIES}
         ${LIBBFD_LIBRARIES}
         ${OPENSSL_LIBRARIES}
         ${LIBXML2_LIBRARIES}
         ${Boost_LIBRARIES}
         )
-target_compile_definitions(libappimage-info PRIVATE ${LIBXML2_DEFINITIONS})
+target_compile_definitions(libappimageinfo PRIVATE ${LIBXML2_DEFINITIONS})
 
-set_target_properties(libappimage-info PROPERTIES VERSION ${LIBAPPIMAGE-INFO_VERSION}
-        SOVERSION ${LIBAPPIMAGE-INFO_VERSION_MAJOR} )
-
-# Prevent CMake from adding the 'lib' prefix to the library.
-set_target_properties(libappimage-info PROPERTIES PREFIX "")
-
-install(TARGETS libappimage-info
+install(TARGETS libappimageinfo
         LIBRARY
         DESTINATION lib
         COMPONENT Libraries
@@ -49,25 +42,26 @@ install(TARGETS libappimage-info
         DESTINATION include
         COMPONENT Development)
 
-set_target_properties(libappimage-info PROPERTIES
+set_target_properties(libappimageinfo PROPERTIES
+        VERSION ${LIBAPPIMAGE-INFO_VERSION} SOVERSION ${LIBAPPIMAGE-INFO_VERSION_MAJOR}
+        PUBLIC_HEADER ${PROJECT_SOURCE_DIR}/include/appimage/appimageinfo.h
+        PREFIX "" # Prevent CMake from adding the 'lib' prefix to the library.
         INSTALL_RPATH_USE_LINK_PATH TRUE
         INSTALL_RPATH "\$ORIGIN/../lib:${CMAKE_INSTALL_PREFIX}/lib")
 
-if (NOT AppImageKit_FOUND)
-    add_dependencies(libappimage-info libappimage)
-endif ()
 
-add_executable(appimage-info app/main.cpp)
-target_link_libraries(appimage-info libappimage-info ${APPIMAGEKIT_LIBRARIES} ${LIBBFD_LIBRARIES})
-target_include_directories(appimage-info PRIVATE ${NLOHMANN_JSON_INCLUDE_DIRS})
-target_compile_definitions(appimage-info PRIVATE PROJECT_NAME="${PROJECT_NAME}" PRIVATE PROJECT_VERSION="${PROJECT_VERSION}")
+add_dependencies(libappimageinfo libappimage libappimageinfo)
 
-set_target_properties(appimage-info PROPERTIES
+add_executable(appimageinfotool app/main.cpp)
+target_link_libraries(appimageinfotool libappimageinfo ${APPIMAGEKIT_LIBRARIES} ${LIBBFD_LIBRARIES})
+target_include_directories(appimageinfotool PRIVATE ${NLOHMANN_JSON_INCLUDE_DIRS})
+target_compile_definitions(appimageinfotool PRIVATE PROJECT_NAME="${PROJECT_NAME}" PRIVATE PROJECT_VERSION="${PROJECT_VERSION}")
+
+set_target_properties(appimageinfotool PROPERTIES
         INSTALL_RPATH_USE_LINK_PATH TRUE
         INSTALL_RPATH "\$ORIGIN/../lib:${CMAKE_INSTALL_PREFIX}/lib")
 
-install(TARGETS appimage-info DESTINATION bin COMPONENT binaries)
-install(FILES app/res/appimage-info.desktop DESTINATION share/applications/ COMPONENT desktop)
-install(FILES app/res/appimage-info.svg DESTINATION share/icons/hicolor/scalable/apps/ COMPONENT desktop)
-
-add_dependencies(appimage-info libappimage-info)
+install(TARGETS appimageinfotool DESTINATION bin COMPONENT binaries)
+install(FILES app/res/appimageinfotool.desktop DESTINATION share/applications/ COMPONENT desktop)
+install(FILES app/res/appimageinfotool.svg DESTINATION share/icons/hicolor/scalable/apps/ COMPONENT desktop)
+add_dependencies(appimageinfotool libappimageinfo)
