+++ src/app/res/appimageinfotool.desktop
@@ -4,8 +4,8 @@ Version=1.0
 Terminal=true
 NoDisplay=true
 Type=Application
-Name=AppImage Info
+Name=AppImage Info Tool
 Comment="Extracts metadata form AppImage files."
-Exec=appimage-info
-Icon=appimage-info
+Exec=appimageinfotool
+Icon=appimageinfotool
 Categories=Utility;
